+++ src/__init__.py
@@ -2,12 +2,11 @@
 
 from .generator import (
     generate_puzzle,
-    save_puzzle,
-    validate_puzzle,
     generate_multiple_puzzles,
-    save_puzzles,
     puzzle_to_ascii,
 )
+from .puzzle_io import save_puzzle, save_puzzles
+from .validator import validate_puzzle
 
 __all__ = [
     "generate_puzzle",
